requirements.txt
Flask==2.3.3
openai==1.41.1
python-dotenv==1.0.0
flask-cors==3.0.10